Update ebay_selenium_webcrawler_product_sale_price.py
Clean code by moving library imports outside of (and before) the main() executable. 

Also, added logic to webcrawler to check if the number of listings that match the keyword number more than 60: i.e., the default number of listings per page. If it's less than 60, then only the first n--ie, the number of  "matched" listings--will be retained. If it's greater than 60, then  the webcrawler will continue to click onto the next page of listings until the final page of listings is reached.

Added additional columns of data to scrape, in addition to data cleaning code to clean these scraped data.