Update ebay_selenium_webcrawler_product_sale_price.py
Parse number of bids data and add this as a new column to a DataFrame and then add this as an additional column of the outputted CSV file. The bids data are not scraped directly due to potential alignment of data issues. 

Instead, the bids data are scraped via a boolean "or" conditional argument for a selenium By.xpath() along with the price_type data since each sold listing on eBay will comprise a Buy it Now, Best Offer, or Auction type, but auction listings will never be listed directly and instead will only show the total number of bids that were used during the selling of the item. The auction bids data are then parsed from the price_type column. The price_type column is also updated such that any listing containing 1 or more bids will be classified as an Auction listing.