Update ebay_selenium_webcrawler_product_sale_price.py
Account for eBay listings that do not have shipping prices listed--i.e., literally denoted as "Shipping not specified". In these instances convert these to empty string, and then filter the columns out of the dataset since we need to know the shipping price in order to calculate the true total price (ie, item price +shipping price).